Split parser impl block parsing
